+++ packages/core/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.1.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/core@5.1.0-rc.1...@ceramicnetwork/core@5.1.0) (2024-02-22)
+
+**Note:** Version bump only for package @ceramicnetwork/core
+
+
+
+
+
+# [5.1.0-rc.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/core@5.1.0-rc.0...@ceramicnetwork/core@5.1.0-rc.1) (2024-02-21)
+
+
+### Bug Fixes
+
+* AnchorProcessingLoop can process the same stream concurrently ([#3163](https://github.com/ceramicnetwork/js-ceramic/issues/3163)) ([ceee8a3](https://github.com/ceramicnetwork/js-ceramic/commit/ceee8a3e1b98293472b6a01220c45981df84ad9d))
+
+
+### Features
+
+* automatically register interest in the metamodel when using recon ([#3157](https://github.com/ceramicnetwork/js-ceramic/issues/3157)) ([242c833](https://github.com/ceramicnetwork/js-ceramic/commit/242c833541c1c6a530dcf39bd2b70fe2f0c4bc2d))
+* create recon feed and consume events ([#3161](https://github.com/ceramicnetwork/js-ceramic/issues/3161)) ([0317921](https://github.com/ceramicnetwork/js-ceramic/commit/0317921e0ee52408bd3d303c4b5dbb53f70d483c))
+* Keep ProcessingLoop full up to the concurrency limit ([#3167](https://github.com/ceramicnetwork/js-ceramic/issues/3167)) ([e486762](https://github.com/ceramicnetwork/js-ceramic/commit/e486762b504a7a38d7b17b10d506cb99ee7ad935))
+* recon put events ([#3143](https://github.com/ceramicnetwork/js-ceramic/issues/3143)) ([7cd8dea](https://github.com/ceramicnetwork/js-ceramic/commit/7cd8deab4d4426a28a8b20504ac7709865e1be3b))
+* turn on cross-node sync test in recon mode ([#3165](https://github.com/ceramicnetwork/js-ceramic/issues/3165)) ([e65c8b7](https://github.com/ceramicnetwork/js-ceramic/commit/e65c8b7af2ebaebe5a97fdec58f1ed3654691aba))
+
+
+
+
+
 # [5.1.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/core@5.0.1-rc.0...@ceramicnetwork/core@5.1.0-rc.0) (2024-02-13)