+++ packages/stream-tile-handler/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.1.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile-handler@5.0.1-rc.1...@ceramicnetwork/stream-tile-handler@5.1.0) (2024-02-22)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile-handler
+
+
+
+
+
+## [5.0.1-rc.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile-handler@5.0.1-rc.0...@ceramicnetwork/stream-tile-handler@5.0.1-rc.1) (2024-02-21)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile-handler
+
+
+
+
+
 ## [5.0.1-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile-handler@5.0.0...@ceramicnetwork/stream-tile-handler@5.0.1-rc.0) (2024-02-08)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-tile-handler